refactor: 버튼 타입 분리
