+++ src/components/common/button/ButtonType.tsx
@@ -1,0 +1,8 @@
+export type ButtonProps = {
+  children: React.ReactNode;
+  styles?: string;
+  onClick?: any;
+  disabled?: boolean;
+  type: 'submit'|'button'|'reset';
+  tabIndex?: number;
+}
